requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
colorlog==6.7.0
flask-cors==4.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0